frontend/flyout: collabs and starting to work on quotas/licenses
